refactor: using contentlayer for projects
