+++ app/(site)/about/page.tsx
@@ -5,7 +5,6 @@ import { ExternalLink } from '@/components/ui/external-link'
 import { experiences } from '@/data/experience'
 import { generateMeta } from '@/lib/meta'
 import { Metadata } from 'next'
-import Image from 'next/image'
 
 export const metadata: Metadata = generateMeta({
 	title: 'About' + ' — Nitin Panwar',
