+++ components/providers.tsx
@@ -5,9 +5,7 @@ import { PropsWithChildren } from 'react'
 import { ReactLenis, useLenis } from '@studio-freight/react-lenis'
 
 export const Providers: React.FC<PropsWithChildren> = ({ children }) => {
-	const lenis = useLenis(({ scroll }) => {
-		// called every scroll
-	})
+	useLenis()
 
 	return (
 		<ReactLenis root>
