+++ content/projects/poly.json
@@ -1,0 +1,24 @@
+{
+	"type": "Project",
+	"cover": "/images/projects/poly/poly.webp",
+	"title": "Poly",
+	"slug": "poly",
+	"role": "Fullstack Developer",
+	"platforms": ["web", "mobile"],
+	"links": [
+		{
+			"name": "Github",
+			"href": "https://github.com/nitintf/poly-client"
+		}
+	],
+	"overview": "Poly is a robust project management tool designed to streamline collaboration and enhance efficiency within teams. Drawing inspiration from Jira, Poly goes beyond conventional project management, offering a user-friendly experience for overseeing entire projects.",
+	"keyFeatures": [
+		"Task Management: Poly simplifies task tracking, enabling teams to break down projects into manageable tasks. Assign, prioritize, and monitor progress effortlessly.",
+		"Collaborative Workspace: Foster real-time collaboration with a centralized workspace where team members can share ideas, updates, and resources seamlessly.",
+		"User-Friendly Interface: With an intuitive and visually appealing interface, Poly ensures that team members can navigate and contribute without a steep learning curve."
+	],
+	"tech": {
+		"frontend": ["React", "GraphQL", "Apollo", "Tailwind", "Vite"],
+		"backend": ["NestJs", "PostgreSQL", "Prisma", "GraphQL", "Resend"]
+	}
+}
