--- contentlayer.config.js
@@ -1,112 +1,0 @@
-import { defineDocumentType, makeSource } from 'contentlayer/source-files';
-import remarkGfm from 'remark-gfm';
-import rehypePrettyCode from 'rehype-pretty-code';
-import rehypeSlug from 'rehype-slug';
-import rehypeAutolinkHeadings from 'rehype-autolink-headings';
-
-
-/** @type {import('contentlayer/source-files').ComputedFields} */
-const computedFields = {
-  slug: {
-    type: 'string',
-    resolve: (doc) => doc._raw.flattenedPath,
-  },
-  tweetIds: {
-    type: 'array',
-    resolve: (doc) => {
-      const tweetMatches = doc.body.raw.match(
-        /<StaticTweet\sid="[0-9]+"\s\/>/g
-      );
-      return tweetMatches?.map((tweet) => tweet.match(/[0-9]+/g)[0]) || [];
-    },
-  },
-  structuredData: {
-    type: 'object',
-    resolve: (doc) => ({
-      '@context': 'https://schema.org',
-      '@type': 'BlogPosting',
-      headline: doc.title,
-      datePublished: doc.publishedAt,
-      dateModified: doc.publishedAt,
-      description: doc.summary,
-      image: doc.image
-        ? `https://nitinp.dev${doc.image}`
-        : `https://nitinp.dev/og?title=${doc.title}`,
-      url: `https://nitinp.dev/note/${doc._raw.flattenedPath}`,
-      author: {
-        '@type': 'Person',
-        name: 'Nitin Panwar',
-      },
-    }),
-  },
-};
-
-
-export const Note = defineDocumentType(() => ({
-  name: 'Note',
-  filePathPattern: `**/*.mdx`,
-  contentType: 'mdx',
-  fields: {
-    title: {
-      type: 'string',
-      required: true,
-    },
-    publishedAt: {
-      type: 'string',
-      required: true,
-    },
-    summary: {
-      type: 'string',
-      required: true,
-    },
-    image: {
-      type: 'string',
-      required: true
-    },
-    length: {
-      options: ['Medium', 'Long', 'Short'],
-      default: 'Medium',
-      type: 'enum',
-    }
-  },
-  computedFields,
-}));
-
-export default makeSource({
-  contentDirPath: 'content',
-
-  documentTypes: [Note],
-  mdx: {
-    remarkPlugins: [remarkGfm],
-    rehypePlugins: [
-      rehypeSlug,
-      [
-        rehypePrettyCode,
-        {
-          theme: 'nord',
-          onVisitLine(node) {
-            // Prevent lines from collapsing in `display: grid` mode, and allow empty
-            // lines to be copy/pasted
-            if (node.children.length === 0) {
-              node.children = [{ type: 'text', value: ' ' }];
-            }
-          },
-          onVisitHighlightedLine(node) {
-            node.properties.className = ['line--highlighted'];
-          },
-          onVisitHighlightedChars(node) {
-            node.properties.className = ['word--highlighted'];
-          },
-        },
-      ],
-      [
-        rehypeAutolinkHeadings,
-        {
-          properties: {
-            className: ['anchor'],
-          },
-        },
-      ],
-    ],
-  },
-});
